pyproject.toml
[tool.poetry]
name = "ai-cto"
version = "0.1.0"
description = ""
authors = ["Germuu <jeremyyamajako@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
flask = "^3.1.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
openai = "^1.59.9"
flask-sqlalchemy = "^3.1.1"
bokeh = "^3.6.2"
robotframework = "^7.2"
pytest = "^8.3.4"
robotframework-seleniumlibrary = "^6.7.0"
webdriver-manager = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
